@@ -1,31 +1,25 @@
 module k8s.io/kube-state-metrics/v2
 
 require (
-	github.com/Azure/go-autorest/autorest v0.10.0 // indirect
-	github.com/brancz/gojsontoyaml v0.0.0-20190425155809-e8bd32d46b3d
+	github.com/brancz/gojsontoyaml v0.0.0-20210709073738-09985366e724
 	github.com/campoy/embedmd v1.0.0
 	github.com/dgryski/go-jump v0.0.0-20170409065014-e1f439676b57
-	github.com/fatih/color v1.9.0 // indirect
-	github.com/google/go-jsonnet v0.14.0
+	github.com/google/go-jsonnet v0.17.0
 	github.com/jsonnet-bundler/jsonnet-bundler v0.4.1-0.20200708074244-ada055a225fa
-	github.com/mattn/go-colorable v0.1.7 // indirect
-	github.com/matttproud/golang_protobuf_extensions v1.0.2-0.20181231171920-c182affec369 // indirect
 	github.com/oklog/run v1.1.0
 	github.com/pkg/errors v0.9.1
-	github.com/prometheus/client_golang v1.7.1
+	github.com/prometheus/client_golang v1.11.0
 	github.com/prometheus/client_model v0.2.0
-	github.com/prometheus/common v0.15.0
-	github.com/prometheus/prometheus v2.5.0+incompatible
-	github.com/robfig/cron/v3 v3.0.0
+	github.com/prometheus/common v0.29.0
+	github.com/prometheus/exporter-toolkit v0.6.0
+	github.com/robfig/cron/v3 v3.0.1
 	github.com/spf13/pflag v1.0.5
-	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d // indirect
-	golang.org/x/sys v0.0.0-20200915084602-288bc346aa39 // indirect
-	golang.org/x/tools v0.0.0-20200616133436-c1934b75d054
-	k8s.io/api v0.19.3
-	k8s.io/apimachinery v0.19.3
-	k8s.io/autoscaler/vertical-pod-autoscaler v0.0.0-20200727194258-b7922d74509c
-	k8s.io/client-go v0.19.3
-	k8s.io/klog/v2 v2.2.0
+	golang.org/x/tools v0.1.4
+	k8s.io/api v0.21.2
+	k8s.io/apimachinery v0.21.2
+	k8s.io/autoscaler/vertical-pod-autoscaler v0.9.2
+	k8s.io/client-go v0.21.2
+	k8s.io/klog/v2 v2.9.0
 )
 
-go 1.14
+go 1.16
