@@ -30,11 +30,12 @@ import (
 )
 
 type generateMetricsTestCase struct {
-	Obj         interface{}
-	MetricNames []string
-	Want        string
-	Headers     []string
-	Func        func(interface{}) []metric.FamilyInterface
+	Obj             interface{}
+	MetricNames     []string
+	AllowLabelsList []string
+	Want            string
+	Headers         []string
+	Func            func(interface{}) []metric.FamilyInterface
 }
 
 func (testCase *generateMetricsTestCase) run() error {
