@@ -48,9 +48,9 @@ type MetricsHandler struct {
 
 	cancel func()
 
-	// mtx protects stores, curShard, and curTotalShards
+	// mtx protects metricsWriters, curShard, and curTotalShards
 	mtx            *sync.RWMutex
-	stores         []cache.Store
+	metricsWriters []metricsstore.MetricsWriter
 	curShard       int32
 	curTotalShards int
 }
@@ -81,7 +81,7 @@ func (m *MetricsHandler) ConfigureSharding(ctx context.Context, shard int32, tot
 	ctx, m.cancel = context.WithCancel(ctx)
 	m.storeBuilder.WithSharding(shard, totalShards)
 	m.storeBuilder.WithContext(ctx)
-	m.stores = m.storeBuilder.Build()
+	m.metricsWriters = m.storeBuilder.Build()
 	m.curShard = shard
 	m.curTotalShards = totalShards
 }
@@ -174,8 +174,8 @@ func (m *MetricsHandler) Run(ctx context.Context) error {
 	return ctx.Err()
 }
 
-// ServeHTTP implements the http.Handler interface. It writes the metrics in
-// its stores to the response body.
+// ServeHTTP implements the http.Handler interface. It writes all generated
+// metrics to the response body.
 func (m *MetricsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	m.mtx.RLock()
 	defer m.mtx.RUnlock()
@@ -198,9 +198,8 @@ func (m *MetricsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	for _, s := range m.stores {
-		ms := s.(*metricsstore.MetricsStore)
-		ms.WriteAll(w)
+	for _, w := range m.metricsWriters {
+		w.WriteAll(writer)
 	}
 
 	// In case we gzipped the response, we have to close the writer.
